Add keras pickleable mixin
