+++ keras/src/models/model.py
@@ -11,6 +11,7 @@ from keras.src.api_export import keras_export
 from keras.src.layers.layer import Layer
 from keras.src.models.variable_mapping import map_trackable_variables
 from keras.src.saving import saving_api
+from keras.src.saving.keras_pickleable import KerasPickleable
 from keras.src.trainers import trainer as base_trainer
 from keras.src.utils import summary_utils
 from keras.src.utils import traceback_utils
@@ -32,7 +33,7 @@ else:
 
 
 @keras_export(["keras.Model", "keras.models.Model"])
-class Model(Trainer, base_trainer.Trainer, Layer):
+class Model(Trainer, base_trainer.Trainer, Layer, KerasPickleable):
     """A model grouping layers into an object with training/inference features.
 
     There are three ways to instantiate a `Model`:
@@ -348,30 +349,6 @@ class Model(Trainer, base_trainer.Trainer, Layer):
         """
         saving_api.load_weights(
             self, filepath, skip_mismatch=skip_mismatch, **kwargs
-        )
-
-    # Note: renaming this function will cause old pickles to be broken.
-    # This is probably not a huge deal, as pickle should not be a recommended
-    # saving format -- it should only be supported for use with distributed
-    # computing frameworks.
-    @classmethod
-    def _unpickle_model(cls, bytesio):
-        # pickle is not safe regardless of what you do.
-        return saving_lib._load_model_from_fileobj(
-            bytesio, custom_objects=None, compile=True, safe_mode=False
-        )
-
-    def __reduce__(self):
-        """__reduce__ is used to customize the behavior of `pickle.pickle()`.
-
-        The method returns a tuple of two elements: a function, and a list of
-        arguments to pass to that function.  In this case we just leverage the
-        keras saving library."""
-        buf = io.BytesIO()
-        saving_lib._save_model_to_fileobj(self, buf, "h5")
-        return (
-            self._unpickle_model,
-            (buf,),
         )
 
     def quantize(self, mode):
