+++ keras/src/optimizers/base_optimizer.py
@@ -6,11 +6,12 @@ from keras.src import initializers
 from keras.src import ops
 from keras.src.optimizers.schedules import learning_rate_schedule
 from keras.src.saving import serialization_lib
+from keras.src.saving.keras_pickleable import KerasPickleable
 from keras.src.utils import tracking
 from keras.src.utils.naming import auto_name
 
 
-class BaseOptimizer:
+class BaseOptimizer(KerasPickleable):
     def __init__(
         self,
         learning_rate,
