+++ keras/src/saving/__init__.py
@@ -5,5 +5,6 @@ from keras.src.saving.object_registration import get_registered_name
 from keras.src.saving.object_registration import get_registered_object
 from keras.src.saving.object_registration import register_keras_serializable
 from keras.src.saving.saving_api import load_model
+from keras.src.saving.keras_pickleable import KerasPickleable
 from keras.src.saving.serialization_lib import deserialize_keras_object
 from keras.src.saving.serialization_lib import serialize_keras_object
